Pots: fix mux cross-bleed by accounting for ADC pipeline depth on flip
Pots::on_adc_sample flipped the mux GPIO synchronously when an averaged
read completed, but the AdcEngine drains samples in batches — so multiple
pre-flip samples (still on the OLD mux state) were dispatched AFTER the
flip and incorrectly accumulated into the NEW pot's average. With the
default samples_per_read = 6, a single stale pre-flip sample bleeding
through produced a `prev_pot_value / 6` contamination term in the new
pot's reading. Easy repro: set pot 0 to min and pot 2 to max — pot 0
bounces bimodally between 0 and ~pot2_value/samples_per_read.

settle_discard_samples (default 2) was sized for the analog mux+S/H
settle time (sub-microsecond), not for the drain pipeline depth, which
is the dominant lag in the 2.1 architecture. Two unrelated phenomena
were being collapsed into one knob.

Fix: AdcEngine exposes pending_sample_count_for_channel_unlocked(),
safely callable from inside a SampleCallback (where the engine's lock
is already held). It returns "samples for this channel currently in
flight" = (unread samples in the DMA ring for that channel) +
(conservative ADC FIFO/in-progress allowance, distributed across active
channels).

Pots calls this at the moment of mux flip and sets pot_discard_remaining_
to (pipeline_pending + settle_discard_samples). The pipeline samples are
discarded automatically before the new pot's accumulator starts filling,
so cross-bleed is impossible regardless of drain batch size, channel
count, or ADC clkdiv. No new app-side configuration. settle_discard_
samples keeps its original meaning (analog settle on truly-fresh
samples).

pot_discard_remaining_ widened from uint8_t to uint16_t — worst-case
pipeline depth can briefly exceed 255 if drains are infrequent (e.g.
only Pots is registered and audio's inline drain isn't running).

Adds pot_cross_bleed_test (menu entry 12): user sets pot 0 to min and
pot 2 to max, the test samples pot 0 raw for 4s and PASSes only if the
observed range is below 100 LSB at 12-bit. The unfixed bug produces a
range of ~1365 LSB; analog jitter alone is < 10 LSB. Threshold cleanly
separates the two states.

+++ brain/include/pots-core.h
@@ -132,7 +132,10 @@ private:
 
 	uint32_t pot_accumulator_ = 0;
 	uint8_t pot_samples_collected_ = 0;
-	uint8_t pot_discard_remaining_ = 0;
+	// Held in a 16-bit field because the post-flip discard count includes the
+	// full ADC + DMA pipeline depth, which can briefly exceed 255 if drains
+	// are infrequent (e.g. only `Pots` registered with no audio inline drain).
+	uint16_t pot_discard_remaining_ = 0;
 	uint8_t active_pot_index_ = 0;
 
 	std::function<void(uint8_t, uint16_t)> on_change_;
